fixes a previous race condition between multiple decode threads
- the fetch tasks were guaranteed to process messages in order, since only a single thread would consume each partition,
  and each partition represents a separate sequential queue
- however, these fetch threads would dump the resulting messages to a single queue, consumed by all decode threads
- after the decode threads read from this single queue, they write back to another shared queue of bundles, which can
  arbitrarily shuffle the intermediate queue since there's no coordination between decoders

The fix here is making each fetch thread write to a separate intermediate queue, maintaining the per-partition ordering.
This does mean there can be many more blocking-queues in memory at once (one per partition), and I didnt readjust the default
size.  However, both the number of partitions and the sizes of queues are small enough in practice that this shoudlnt be
a problem.